src/requirements.txt
# Core FastAPI and web framework dependencies
fastapi>=0.109.0,<0.111.0
uvicorn[standard]==0.24.0
pydantic>=2.0.0,<3.0.0
email-validator==2.1.0

# Database and storage
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client and requests
requests==2.31.0
httpx==0.25.2

# Google APIs and integrations
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
google-api-python-client==2.108.0
google-cloud-speech==2.20.0
google-generativeai==0.7.2
firebase-admin==6.3.0

# Communication integrations
twilio==8.10.0
sendgrid==6.10.0

# Payment processing
stripe==7.8.0

# Scheduling and calendar
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development and debugging
python-dotenv==1.0.0
rich==13.7.0

# Async support
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0

# Background tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0 